♻️ Extract Hyrax::GroupBehavior
As of 2023-11-09, [Hyrax::Group in Hyku's main branch][1] inherits
from ApplicationRecord.  This means to share logic between the two
requires that we extract a module (as we cannot change inheritance).

This commit is that effort.

[1]: https://github.com/samvera/hyku/blob/dd0da8061c2ee0233e4f32f8ea7748eda2792ba3/app/models/hyrax/group.rb#L6
